Refactored the constraints delegation logic
Modified how constraints delegation are stored in leader data. Instead
of storing a marshaled array with all the computers that have the
constraints set, we set individual leader variables. This will avoid
the scenario when the marshaled array is big enough and might cause the
leader-set.exe call to break.